+++ src/mwax_mover/mwacache_archive_processor.py
@@ -38,10 +38,11 @@ class MWACacheArchiveProcessor:
     """
 
     def __init__(self):
-        self.logger: logging.Logger
+        self.logger = logging.getLogger(__name__)
 
         self.hostname: str = ""
         self.log_path: str = ""
+        self.log_level: str = ""
         self.metafits_path: str = ""
         self.archive_to_location: ArchiveLocation = ArchiveLocation.Unknown
         self.concurrent_archive_workers: int = 0
@@ -446,20 +447,23 @@ class MWACacheArchiveProcessor:
             print(f"log_path {self.log_path} does not exist. Quiting.")
             sys.exit(1)
 
+        # Read log level
+        config_file_log_level: Optional[str] = utils.read_optional_config(
+            self.logger, config, "mwax mover", "log_level"
+        )
+        if config_file_log_level is None:
+            self.log_level = "DEBUG"
+            self.logger.warning(f"log_level not set in config file. Defaulting to {self.log_level} level logging.")
+        else:
+            self.log_level = config_file_log_level
+
         # It's now safe to start logging
         # start logging
-        self.logger = logging.getLogger(__name__)
-        self.logger.setLevel(logging.DEBUG)
+        self.logger.setLevel(self.log_level)
         console_log = logging.StreamHandler()
-        console_log.setLevel(logging.DEBUG)
+        console_log.setLevel(self.log_level)
         console_log.setFormatter(logging.Formatter("%(asctime)s, %(levelname)s, %(threadName)s, %(message)s"))
         self.logger.addHandler(console_log)
-
-        # Removing file logging for now
-        # file_log = logging.FileHandler(filename=os.path.join(self.log_path, "mwacache_archiver_main.log"))
-        # file_log.setLevel(logging.DEBUG)
-        # file_log.setFormatter(logging.Formatter("%(asctime)s, %(levelname)s, %(threadName)s, %(message)s"))
-        # self.logger.addHandler(file_log)
 
         self.logger.info(
             "Starting mwacache_archive_processor" f" processor...v{version.get_mwax_mover_version_string()}"