+++ tests/mwax_subfile_distributor_correlator_test.py
@@ -82,8 +82,11 @@ def setup_test_dirs():
     metafits_path = os.path.join(base_dir, "vulcan_metafits")
     check_and_make_dir(metafits_path)
 
-    packet_stats_dump_dir = os.path.join("packet_stats_dump_dir")
+    packet_stats_dump_dir = os.path.join(base_dir, "packet_stats_dump_dir")
     check_and_make_dir(packet_stats_dump_dir)
+
+    packet_stats_destination_dir = os.path.join(base_dir, "packet_stats_destination_dir")
+    check_and_make_dir(packet_stats_destination_dir)
 
 
 def test_correlator_config_file():